[issue-407, refactor] introduce module "rdfschema" to use common spdx namespaces in rdf writer and parser
