+++ src/spdx/rdfschema/namespace.py
@@ -1,0 +1,14 @@
+# Copyright (c) 2023 spdx contributors
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#     http://www.apache.org/licenses/LICENSE-2.0
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from rdflib import Namespace
+
+SPDX_NAMESPACE = Namespace("http://spdx.org/rdf/terms#")
+POINTER_NAMESPACE = Namespace("http://www.w3.org/2009/pointers#")
