+++ src/spdx/writer/rdf/annotation_writer.py
@@ -15,7 +15,8 @@ from rdflib import Graph, Literal, RDFS, URIRef, RDF, BNode
 from spdx.datetime_conversions import datetime_to_iso_string
 from spdx.model.annotation import Annotation
 from spdx.writer.casing_tools import snake_case_to_camel_case
-from spdx.writer.rdf.writer_utils import SPDX_NAMESPACE, add_namespace_to_spdx_id
+from spdx.writer.rdf.writer_utils import add_namespace_to_spdx_id
+from spdx.rdfschema.namespace import SPDX_NAMESPACE
 
 
 def add_annotation_to_graph(annotation: Annotation, graph: Graph, doc_namespace: str,
