+++ src/spdx/writer/rdf/creation_info_writer.py
@@ -13,7 +13,8 @@ from rdflib import Graph, BNode, RDF, Literal, RDFS, URIRef
 from spdx.datetime_conversions import datetime_to_iso_string
 from spdx.model.document import CreationInfo
 from spdx.writer.rdf.external_document_ref_writer import add_external_document_ref_to_graph
-from spdx.writer.rdf.writer_utils import SPDX_NAMESPACE, add_optional_literal
+from spdx.writer.rdf.writer_utils import add_optional_literal
+from spdx.rdfschema.namespace import SPDX_NAMESPACE
 
 
 def add_creation_info_to_graph(creation_info: CreationInfo, graph: Graph):
