+++ src/spdx/writer/rdf/package_writer.py
@@ -18,8 +18,9 @@ from spdx.writer.rdf.checksum_writer import add_checksum_to_graph
 
 from spdx.model.package import Package, PackageVerificationCode, ExternalPackageRef, \
     CATEGORY_TO_EXTERNAL_PACKAGE_REF_TYPES
-from spdx.writer.rdf.writer_utils import SPDX_NAMESPACE, add_optional_literal, add_literal_or_no_assertion_or_none, \
+from spdx.writer.rdf.writer_utils import add_optional_literal, add_literal_or_no_assertion_or_none, \
     add_datetime_to_graph, add_namespace_to_spdx_id
+from spdx.rdfschema.namespace import SPDX_NAMESPACE
 
 
 def add_package_to_graph(package: Package, graph: Graph, doc_namespace: str,
