+++ src/spdx/writer/rdf/relationship_writer.py
@@ -16,7 +16,8 @@ from spdx.model.relationship import Relationship
 from spdx.model.spdx_no_assertion import SpdxNoAssertion
 from spdx.model.spdx_none import SpdxNone
 from spdx.writer.casing_tools import snake_case_to_camel_case
-from spdx.writer.rdf.writer_utils import SPDX_NAMESPACE, add_namespace_to_spdx_id
+from spdx.writer.rdf.writer_utils import add_namespace_to_spdx_id
+from spdx.rdfschema.namespace import SPDX_NAMESPACE
 
 
 def add_relationship_to_graph(relationship: Relationship, graph: Graph, doc_namespace: str,
