+++ tests/spdx/writer/rdf/test_checksum_writer.py
@@ -13,7 +13,7 @@ from rdflib import Graph, URIRef, Literal
 
 from spdx.model.checksum import ChecksumAlgorithm
 from spdx.writer.rdf.checksum_writer import add_checksum_to_graph, algorithm_to_rdf_string
-from spdx.writer.rdf.writer_utils import SPDX_NAMESPACE
+from spdx.rdfschema.namespace import SPDX_NAMESPACE
 from tests.spdx.fixtures import checksum_fixture
