+++ tests/spdx/writer/rdf/test_extracted_licensing_info_writer.py
@@ -9,7 +9,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from rdflib import Graph, Literal, RDFS, URIRef
-from spdx.writer.rdf.writer_utils import SPDX_NAMESPACE
+from spdx.rdfschema.namespace import SPDX_NAMESPACE
 
 from spdx.writer.rdf.extracted_licensing_info_writer import add_extracted_licensing_info_to_graph
 from tests.spdx.fixtures import extracted_licensing_info_fixture
